Move redirect to separate package
