+++ src/test/java/com/flowingcode/vaadin/addons/ironicons/redirect/DemoViewSimpleRedirect.java
@@ -1,0 +1,37 @@
+/*-
+ * #%L
+ * Iron Icons
+ * %%
+ * Copyright (C) 2019 Flowing Code
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+ package com.flowingcode.vaadin.addons.ironicons.redirect;
+
+import com.vaadin.flow.component.orderedlayout.VerticalLayout;
+import com.vaadin.flow.router.BeforeEnterEvent;
+import com.vaadin.flow.router.BeforeEnterObserver;
+import com.vaadin.flow.router.Route;
+
+import com.flowingcode.vaadin.addons.ironicons.DemoViewSingle;
+
+@Route("view")
+@SuppressWarnings("serial")
+public class DemoViewSimpleRedirect extends VerticalLayout implements BeforeEnterObserver {
+
+	@Override
+	public void beforeEnter(BeforeEnterEvent event) {
+		event.forwardTo(DemoViewSingle.class);
+	}
+}
